services/api-gateway/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
langchain==0.2.2
langchain-community==0.2.1
langchain-openai==0.1.8
qdrant-client==1.9.0
psycopg2-binary==2.9.9
pydantic==2.7.1
loguru==0.7.2
python-dotenv==1.0.1
openai==1.30.1
pyyaml==6.0.1
pypdf==4.2.0
python-docx==1.1.2
